Fix prefix usage and update tests
